ai/requirements.txt
fastapi==0.116.1
langchain==0.3.27
langchain_community==0.3.27
langchain_huggingface==0.3.1
openai==1.101.0
pydantic==2.11.7
PyYAML==6.0.2
PyYAML==6.0.2
streamlit==1.46.1
torch==2.7.1+cu126
transformers==4.55.3
